requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2

# Time Series Forecasting (Prophet is optional - see installation notes below)
# prophet==1.1.4  # Uncomment if you can install Prophet successfully
statsmodels==0.14.0

# Dashboard and Visualization
streamlit==1.28.1
plotly==5.17.0
altair==5.1.2

# Data Processing
pyyaml==6.0.1
python-dateutil==2.8.2

# Utilities
tqdm==4.66.1
joblib==1.3.2

# Development
jupyter==1.0.0
ipykernel==6.25.2

# Alternative time series libraries (if Prophet fails)
# Uncomment these if you have issues with Prophet:
# pmdarima==2.0.4
# arch==6.2.0 